Assignment Evaluator Solved
